simple_to_do/requirements.txt
streamlit==1.20.0
streamlit-option-menu==0.4.3
pycryptodome==3.14.1
